refactor: ddd 전략을 고수하는 domain 방식 디렉토리  폴더 생성
