+++ backend/footprint/src/main/java/foot/footprint/domain2/Article.java
@@ -1,4 +1,4 @@
-package foot.footprint.domain;
+package foot.footprint.domain2;
 
 import lombok.Builder;
 import lombok.Getter;
