+++ backend/footprint/src/main/java/foot/footprint/domain2/typeHandler/CodeEnum.java
@@ -1,4 +1,4 @@
-package foot.footprint.domain.typeHandler;
+package foot.footprint.domain2.typeHandler;
 
 public interface CodeEnum {
     String getCode();
