+++ backend/footprint/src/main/java/foot/footprint/domain2/user/Role.java
@@ -1,8 +1,8 @@
-package foot.footprint.domain.user;
+package foot.footprint.domain2.user;
 
 
-import foot.footprint.domain.typeHandler.CodeEnum;
-import foot.footprint.domain.typeHandler.CodeEnumTypeHandler;
+import foot.footprint.domain2.typeHandler.CodeEnum;
+import foot.footprint.domain2.typeHandler.CodeEnumTypeHandler;
 import org.apache.ibatis.type.MappedTypes;
 
 public enum Role implements CodeEnum {
