+++ backend/footprint/src/main/java/foot/footprint/global/error/ExceptionAdvice.java
@@ -1,0 +1,4 @@
+package foot.footprint.global.error;
+
+public class ExceptionAdvice {
+}
