+++ backend/footprint/src/main/resources/mybatis-config.xml
@@ -11,8 +11,8 @@
     </settings>
 
     <typeHandlers>
-        <typeHandler handler="foot.footprint.domain.user.Role$TypeHandler"/>
-        <typeHandler handler="foot.footprint.domain.user.AuthProvider$TypeHandler"/>
-        <typeHandler handler="foot.footprint.domain.typeHandler.CodeEnumTypeHandler"/>
+        <typeHandler handler="foot.footprint.domain2.user.Role$TypeHandler"/>
+        <typeHandler handler="foot.footprint.domain2.user.AuthProvider$TypeHandler"/>
+        <typeHandler handler="foot.footprint.domain2.typeHandler.CodeEnumTypeHandler"/>
     </typeHandlers>
 </configuration>
