+++ backend/footprint/src/test/java/foot/footprint/repository/article/FIndArticleRepositoryTest.java
@@ -1,6 +1,6 @@
 package foot.footprint.repository.article;
 
-import foot.footprint.domain.Article;
+import foot.footprint.domain2.Article;
 import foot.footprint.repository.RepositoryTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
