+++ backend/footprint/src/test/java/foot/footprint/repository/user/UserRepositoryTest.java
@@ -1,8 +1,8 @@
 package foot.footprint.repository.user;
 
-import foot.footprint.domain.user.AuthProvider;
-import foot.footprint.domain.user.Role;
-import foot.footprint.domain.user.User;
+import foot.footprint.domain2.user.AuthProvider;
+import foot.footprint.domain2.user.Role;
+import foot.footprint.domain2.user.User;
 import foot.footprint.repository.RepositoryTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
